Set max length for username and password on the client.
My goal is a minimal yet secure implementation.

Twitter restricts usernames to 15 characters, so that seems reasonable for this
little app.

NIST and OWASP both recommend maximum password lengths of at least 64
characters. There are benefits to longer passwords, but that also risks the
possibility of denial of service attacks.

- https://cheatsheetseries.owasp.org/cheatsheets/Authentication_Cheat_Sheet.html
- https://pages.nist.gov/800-63-3/sp800-63b.html