hashPassword(): Test hash for entropy and randomness.
I'm only iterating thrice, because the hash algorithm iterates 210,000 times,
and I want to be able to run the test suite quickly (< 1 second on my 10 year
old dual core laptop with 16 GB of RAM).

I'm keeping the weak password the same each time to guarantee that we're
generating different hashes, thanks to the random salt. Otherwise anyone with
access to the database would be able to see which users have the same password.